[NFLX] Fix ClusterConnectionState from caching stale host address.
ClusterConnectionState is used by NetworkClient when Consumers try to
connect to the brokers. However, if a broker is terminated and replaced
by another instance, broker ID stays the same but the address is
changed.

Previously, ClusterConnectionState would always return the same host
address given a node ID, from its internal cache. This change will force
the cache to update the host address if the hostname of the node has
changed.